infrastructure/requirements.txt
# SIGINT CDK Requirements
aws-cdk-lib>=2.120.0
constructs>=10.0.0
